Breakup ABYBackend into two files.
