+++ compiler/src/main/kotlin/edu/cornell/cs/apl/viaduct/backend/aby/ABYBackend.kt
@@ -1,11 +1,17 @@
 package edu.cornell.cs.apl.viaduct.backend.aby
 
 import de.tu_darmstadt.cs.encrypto.aby.ABYParty
-import de.tu_darmstadt.cs.encrypto.aby.Circuit
 import de.tu_darmstadt.cs.encrypto.aby.Role
 import de.tu_darmstadt.cs.encrypto.aby.Share
 import de.tu_darmstadt.cs.encrypto.aby.SharingType
 import edu.cornell.cs.apl.viaduct.analysis.TypeAnalysis
+import edu.cornell.cs.apl.viaduct.backend.ABYCircuitBuilder
+import edu.cornell.cs.apl.viaduct.backend.ABYCircuitGate
+import edu.cornell.cs.apl.viaduct.backend.ABYConstGate
+import edu.cornell.cs.apl.viaduct.backend.ABYDummyInGate
+import edu.cornell.cs.apl.viaduct.backend.ABYInGate
+import edu.cornell.cs.apl.viaduct.backend.ABYOperation
+import edu.cornell.cs.apl.viaduct.backend.ABYOperationGate
 import edu.cornell.cs.apl.viaduct.backend.AbstractBackendInterpreter
 import edu.cornell.cs.apl.viaduct.backend.HostAddress
 import edu.cornell.cs.apl.viaduct.backend.LoopBreakSignal
@@ -64,10 +70,10 @@ import edu.cornell.cs.apl.viaduct.syntax.types.ValueType
 import edu.cornell.cs.apl.viaduct.syntax.values.BooleanValue
 import edu.cornell.cs.apl.viaduct.syntax.values.IntegerValue
 import edu.cornell.cs.apl.viaduct.syntax.values.Value
-import kotlinx.collections.immutable.PersistentMap
-import kotlinx.collections.immutable.persistentMapOf
 import java.util.SortedSet
 import java.util.Stack
+import kotlinx.collections.immutable.PersistentMap
+import kotlinx.collections.immutable.persistentMapOf
 
 /** Backend for the ABY MPC framework. */
 class ABYBackend : ProtocolBackend {
@@ -79,8 +85,6 @@ class ABYBackend : ProtocolBackend {
     private var role: Role? = null
 
     override fun initialize(connectionMap: Map<Host, HostAddress>, projection: ProtocolProjection) {
-        System.loadLibrary("ViaductABY")
-
         val protocolHosts: Set<Host> = projection.protocol.hosts
         assert(protocolHosts.size == 2)
 
@@ -121,88 +125,6 @@ class ABYBackend : ProtocolBackend {
             throw Exception("Could not initialize ABY backend")
         }
     }
-}
-
-class ABYCircuitBuilder(
-    val circuit: Circuit,
-    val bitlen: Long,
-    val role: Role
-)
-
-enum class ABYOperation {
-    SUB_GATE {
-        override val numOperands = 2
-        override fun buildABYGate(circuit: Circuit, children: List<Share>): Share =
-            circuit.putSUBGate(children[0], children[1])
-    },
-    ADD_GATE {
-        override val numOperands = 2
-        override fun buildABYGate(circuit: Circuit, children: List<Share>): Share =
-            circuit.putADDGate(children[0], children[1])
-    },
-    MUL_GATE {
-        override val numOperands = 2
-        override fun buildABYGate(circuit: Circuit, children: List<Share>): Share =
-            circuit.putMULGate(children[0], children[1])
-    },
-    AND_GATE {
-        override val numOperands = 2
-        override fun buildABYGate(circuit: Circuit, children: List<Share>): Share =
-            circuit.putANDGate(children[0], children[1])
-    },
-
-    /*
-    INV_GATE {
-        override val numOperands = 1
-        override fun buildABYGate(circuit: Circuit, children: List<Share>): Share =
-            circuit.putINVGate(children[0].)
-    },
-    */
-    GT_GATE {
-        override val numOperands = 2
-        override fun buildABYGate(circuit: Circuit, children: List<Share>): Share =
-            circuit.putGTGate(children[0], children[1])
-    },
-    EQ_GATE {
-        override val numOperands = 2
-        override fun buildABYGate(circuit: Circuit, children: List<Share>): Share =
-            circuit.putEQGate(children[0], children[1])
-    },
-    MUX_GATE {
-        override val numOperands = 3
-        override fun buildABYGate(circuit: Circuit, children: List<Share>): Share =
-            circuit.putMUXGate(children[0], children[1], children[2])
-    };
-
-    abstract val numOperands: Int
-    abstract fun buildABYGate(circuit: Circuit, children: List<Share>): Share
-}
-
-sealed class ABYCircuitGate(val children: List<ABYCircuitGate>) {
-    abstract fun buildABYGate(builder: ABYCircuitBuilder, children: List<Share>): Share
-}
-
-class ABYInGate(val value: Int) : ABYCircuitGate(listOf()) {
-    override fun buildABYGate(builder: ABYCircuitBuilder, children: List<Share>): Share =
-        builder.circuit.putINGate(value.toBigInteger(), builder.bitlen, builder.role)
-}
-
-class ABYDummyInGate : ABYCircuitGate(listOf()) {
-    override fun buildABYGate(builder: ABYCircuitBuilder, children: List<Share>): Share =
-        builder.circuit.putDummyINGate(builder.bitlen)
-}
-
-class ABYConstGate(val value: Int) : ABYCircuitGate(listOf()) {
-    override fun buildABYGate(builder: ABYCircuitBuilder, children: List<Share>): Share =
-        builder.circuit.putCONSGate(value.toBigInteger(), builder.bitlen)
-}
-
-class ABYOperationGate(
-    val operation: ABYOperation,
-    operands: List<ABYCircuitGate>
-) : ABYCircuitGate(operands) {
-    override fun buildABYGate(builder: ABYCircuitBuilder, children: List<Share>): Share =
-        operation.buildABYGate(builder.circuit, children)
 }
 
 private class ABYInterpreter(
@@ -558,7 +480,7 @@ private class ABYInterpreter(
                     ABYOperation.EQ_GATE,
                     listOf(
                         arguments[0],
-                        arguments[1],
+                        arguments[1]
                     )
                 )
 
@@ -597,7 +519,7 @@ private class ABYInterpreter(
         abstract suspend fun update(update: UpdateNameNode, arguments: List<AtomicExpressionNode>)
     }
 
-    inner class ABYImmutableCellObject(var gate: ABYCircuitGate) : ABYClassObject() {
+    class ABYImmutableCellObject(var gate: ABYCircuitGate) : ABYClassObject() {
         override suspend fun query(query: QueryNameNode, arguments: List<AtomicExpressionNode>): ABYCircuitGate {
             return when (query.value) {
                 is Get -> gate
